+++ wxt.config.ts
@@ -8,7 +8,7 @@ export default defineConfig({
   manifest: {
     name: 'Alfred - Shopify Theme Detector',
     description: 'Instantly detect themes on any Shopify store. Streamline your workflow with smart shortcuts and Shopify productivity tools.',
-    version: '2026.03.15',
+    version: '2026.03.15.1',
     action: {
       default_title: 'Alfred',
     },
@@ -39,9 +39,10 @@ export default defineConfig({
       '--disable-features=BlockThirdPartyCookies', // Allows third-party cookies
       // 'chrome-extension://lepphhjodhfojboecomikglfppimdkmj/options.html',
       // 'https://uptek.com',
-      'https://junaid-workspace.myshopify.com/',
+      // 'https://junaid-workspace.myshopify.com/',
       // 'https://partners.shopify.com/1750954/stores/new?store_type=managed_store',
-      // 'https://admin.shopify.com/store/junaid-workspace/themes'
+      // 'https://admin.shopify.com/store/junaid-workspace/themes',
+      'https://partners.shopify.com/1750954/stores/new?store_domain=zeeshan-h-bhatti.myshopify.com&store_type=managed_store',
     ],
   },
 });